removed a warning
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@12591 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_statmech/statmech_manager.cpp
@@ -3843,7 +3843,7 @@ std::vector<int> StatMechManager::Permutation(const int& N)
 	for (int i=0; i<N; ++i)
 	{
 		//generate random number between 0 and i
-		j = floor((i + 1.0)*uniformclosedopengen_.random());
+		j = (int)floor((i + 1.0)*uniformclosedopengen_.random());
 
 		/*exchange values at positions i and j (note: value at position i is i due to above initialization
 		 *and because so far only positions <=i have been changed*/
